+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.0.4
-## XX/XX/2017
+## 08/16/2017
 
 1. [](#new)
     * CLI: Added `status` command to check config and git (#52, thanks @karfau)